requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
torch>=1.9.0
torchvision>=0.10.0
diffusers>=0.21.0
Pillow>=8.0.0
numpy>=1.21.0
tqdm>=4.60.0
